+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Ep 3
+
+- Using the HTTP methods with express.
+- Using course example to use HTTP examples
+
 ## Ep 2
 
 - Add nodemon to refresh api build everi time if has some changes.